refactor project into its own app
