+++ project/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from project.models import Project
+
+admin.site.register(Project)
