+++ user_profile/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import Skill, Project, Profile
+from .models import Skill, Profile
 
 admin.site.register(Skill)
 admin.site.register(Profile)
-admin.site.register(Project)
